Parse response score as Double instead of Float to resolve permanent loading (#29)
* Parse response score as Double instead of Float to resolve app remaining permanently loading

* rename iam_api_key to iam_apikey as per documentation:
https://console.bluemix.net/apidocs/visual-recognition?language=python

* add get_result() as per documentation:
https://console.bluemix.net/apidocs/visual-recognition?language=python#classify-images

* shorten line under 79 characters to follow the flake8 guideline

* update as recommended To follow the flake8 guideline

* Update run.py

+++ mobile-app/WatsonWasteSorter/ViewController.swift
@@ -53,7 +53,7 @@ class ViewController: UIViewController, UIImagePickerControllerDelegate, UINavig
                                         let JSON = result as! NSDictionary
                                         print(JSON)
                                         if let res = JSON["result"] as? String {
-                                            if let confs = JSON["confident score"] as? Float {
+                                            if let confs = JSON["confident score"] as? Double {
                                                 self.showAlert(text: res, conf:Int(confs * 100))
                                             }
                                         }
